# openapi.yaml
openapi: 3.0.3
info:
  title: Blockvisor API
  description: |
    Blockvisor API is a distributed infrastructure management system that orchestrates software deployment and monitoring across physical and virtual machines. Originally designed for blockchain node management, it has evolved into a general-purpose platform for running and managing any software.
    
    ## Core Concepts
    - **Hosts**: Physical or virtual machines registered with the system via the `bvup` tool
    - **Nodes**: Running software instances with real-time state and metrics monitoring
    - **Organizations**: Groups of users sharing resources and permissions
    - **Users**: Individual accounts with role-based access control (RBAC)
    - **Commands**: Instructions sent to hosts for software lifecycle management
    
    ## Authentication
    The API uses JWT tokens for authentication. Include the token in the `Authorization` header as `Bearer <token>`.
    
    ## Error Handling
    The API returns structured error responses with detailed parameter validation information:
    ```json
    {
      "error": "Invalid query parameters",
      "details": [
        {
          "parameter": "org_id",
          "error": "Invalid UUID format: 'not-a-uuid'",
          "expected": "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"
        }
      ]
    }
    ```
  version: 0.4.0
  contact:
    name: Blockvisor API Support
  license:
    name: MIT
servers:
  - url: https://api.blockvisor.com
...
security:
  - BearerAuth: []
paths:
...
  /v1/user:
    get:
      tags:
        - Users
      summary: List users
      description: Get a list of users with optional filtering
      parameters:
        - name: user_id
          in: query
          description: User ID to filter by (singular form)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: user_ids
          in: query
          description: User IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        - name: org_id
          in: query
          description: Organization ID to filter by (singular form)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: org_ids
          in: query
          description: Organization IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
            example: 50
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                    description: Total number of users
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
  /v1/user/{user_id}:
    get:
      tags:
        - Users
      summary: Get user
      description: Get details of a specific user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"

  # Protocol endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid query parameters"
        details:
          type: array
          items:
            type: object
            properties:
              parameter:
                type: string
                example: "org_id"
              error:
                type: string
                example: "Invalid UUID format: 'not-a-uuid'"
              expected:
                type: string
                example: "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"

    # Core entity schemas
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...
    CreateUserRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        password:
          type: string
          format: password
          example: "securepassword123"
...